Relocate `Logger` to `logging.go`
